test(web): isolate MCP binding lookup test
